Pull request fixes
**My Task Tab Filters**
- The My Tasks tab filters can now be configured in the same way the custom tabs can. This replaces the previous feature commit where excluded statuses could be added in the config.
- The resolve_filters method was moved from the `file_form_base.py`’s `FileFormBase._build_entity_models` method to `util.py`, so that it could be shared with the `my_tasks_model.py`
- The info.yml has been updated to allow filters to be specified for the My Tasks tab. However this is not as clean a I would like. I could not get it to have a list of list, due to validation issues. So it has a dictionary of lists instead.

**Other Fixes**
- Tooltip has been added to the search box to inform users they must press enter to complete the search.
- when getting auto expand setting from config, it no longer specifies a default value, and relies on a value being passed from config or info.yml
