+++ python/tk_multi_workfiles/util.py
@@ -261,3 +261,28 @@ def get_template_user_keys(template):
         if key.shotgun_entity_type == "HumanUser":
             user_keys.add(key.name)
     return user_keys
+
+
+def resolve_filters(filters):
+    app = sgtk.platform.current_bundle()
+
+    resolved_filters = []
+    for filter in filters:
+        if type(filter) is dict:
+            resolved_filter = {
+                "filter_operator": filter["filter_operator"],
+                "filters": resolve_filters(filter["filters"])}
+        else:
+            resolved_filter = []
+            for field in filter:
+                if field == "{context.entity}":
+                    field = app.context.entity
+                elif field == "{context.step}":
+                    field = app.context.step
+                elif field == "{context.task}":
+                    field = app.context.task
+                elif field == "{context.user}":
+                    field = app.context.user
+                resolved_filter.append(field)
+        resolved_filters.append(resolved_filter)
+    return resolved_filters
